+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Add default `title` and `providerIdToAssign` arguments to `createConversation()` endpoint. Fixes [#1](https://github.com/nabla/nabla-ios/issues/1).
+
 
 ## [1.0.0-alpha06] - 2022-06-21